WebKit2 is leaking NSCursors created by leakNamedCursor https://bugs.webkit.org/show_bug.cgi?id=64241 <rdar://problem/9507151>
Reviewed by Oliver Hunt.

../WebCore: 

* platform/mac/CursorMac.mm:
(WebCore::createNamedCursor):
Rename this from leakNamedCursor to createNamedCursor and make it return a
RetainPtr<NSCursor> instead of a raw pointer.

(WebCore::Cursor::ensurePlatformCursor):
Don't leak cursors here. We still won't deallocate cursors during shutdown (which leakNamedCursor
was said to prevent) because the cursor singletons are all allocated from the heap and are never destroyed
anyway.

../WebKit2: 

* Shared/WebCoreArgumentCoders.cpp:
(CoreIPC::::decode):
When decoding a cursor of a known type, make sure to eagerly create the platform cursor
for the cursor singleton. This way we avoid re-creating new NSCursor objects over and over for
standard cursors.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@90705 268f45cc-cd09-0410-ab3c-d52691b4dbfc