Move HttpLock to using a FileSystem object https://bugs.webkit.org/show_bug.cgi?id=64885
Reviewed by Adam Barth.

I made a typo in my previous change, taking the value
of read_text_file(pid_file) and setting it to lock_pid_file
instead of current_pid.  Fixed now. :)

In order to test my new change I had to overhaul the unittests
for this class to create a separate set of tests which work off
of Mock objects instead of the real filesystem.

Since Executive doesn't yet wrap os.getpid() I added a FIXME
in several places where we're currently calling os.getpid().  I
felt adding Executive.getpid was outside of the scope of this change
but once it exists some of this code will be much simpler to mock.

* Scripts/webkitpy/layout_tests/port/http_lock.py:
* Scripts/webkitpy/layout_tests/port/http_lock_unittest.py:
* Scripts/webkitpy/tool/mocktool.py:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@91411 268f45cc-cd09-0410-ab3c-d52691b4dbfc