Date.prototype.toISOString doesn't handle negative years or years > 9999 correctly. https://bugs.webkit.org/show_bug.cgi?id=63986
Patch by Mark Hahnenberg <mhahnenberg@apple.com> on 2011-07-22
Reviewed by Geoffrey Garen.

Source/JavaScriptCore:

Changed the implementation of Date.prototype.toISOString() to use the extended year
format (+/-yyyyyy) for years outside of [0,9999] to be in compliance with ES 15.9.1.15.1.

* runtime/DatePrototype.cpp:
(JSC::dateProtoFuncToISOString):

LayoutTests:

Changed the implementation of Date.prototype.toISOString() to use the extended year
format (+/-yyyyyy) for years outside of [0,9999] to be in compliance with ES 15.9.1.15.1.
Altered one test that tested for incorrect behavior on years > 9999, and added another test for
years < 0 (i.e. BCE).

* fast/js/date-toisostring-expected.txt:
* fast/js/script-tests/date-toisostring.js:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@91576 268f45cc-cd09-0410-ab3c-d52691b4dbfc