[chromium] Reorder functions in LayerRendererChromium.cpp to make bug 58840 easier to review https://bugs.webkit.org/show_bug.cgi?id=65354
Reviewed by Kenneth Russell.

This moves a few function definitions in LayerRendererChromium around to make the patch on bug 58840 easier to
review.  In that patch, updatePropertiesAndRenderSurfaces() becomes a templated free function instead of a
member function, so it has to be near the top of the .cpp.

* platform/graphics/chromium/LayerRendererChromium.cpp:
(WebCore::calculateVisibleRect):
(WebCore::isScaleOrTranslation):
(WebCore::LayerRendererChromium::updatePropertiesAndRenderSurfaces):
(WebCore::LayerRendererChromium::updateLayers):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@92037 268f45cc-cd09-0410-ab3c-d52691b4dbfc