Unreviewed Chromium Windows build fix after r93140. Take 2.
This just makes the troublesome enum public. If this
fixes the build I will file a follow-up bug to clean-up
these build issues and either keep the enum public and
cleanup other functions or restore the enum to private.
There is no harm in making the enum public right now.

* platform/graphics/FontCache.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@93151 268f45cc-cd09-0410-ab3c-d52691b4dbfc