Chromium Mac: Show scrollbar when doing search https://bugs.webkit.org/show_bug.cgi?id=66209
Patch by Sailesh Agrawal <sail@chromium.org> on 2011-08-18
Reviewed by James Robinson.

Source/WebCore:

Updated the Lion scrollbar drawing code to force the scrollbar to be visible when we have tickmarks for search results.

No tests because there are no Mac 10.7 chromium bots yet.

* platform/chromium/ScrollbarThemeChromiumMac.h:
* platform/chromium/ScrollbarThemeChromiumMac.mm:
(WebCore::ScrollbarThemeChromiumMac::paint):
(WebCore::ScrollbarThemeChromiumMac::paintTickmarks):

Source/WebKit/chromium:

Fixed a bug in invalidateArea where it wouldn't properly invalidate overlay scrollbars. The problem was that it was assuming that the scrollbar was always to the right of the visible scrollbar. This isn't true for overlay scrollbars.

No tests because there are no Mac 10.7 chromium bots yet.

* src/WebFrameImpl.cpp:
(WebKit::WebFrameImpl::invalidateArea):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@93361 268f45cc-cd09-0410-ab3c-d52691b4dbfc