Crashes in WebCore::InsertListCommand::unlistifyParagraph. https://bugs.webkit.org/show_bug.cgi?id=67918
Patch by Shinya Kawanaka <shinyak@google.com> on 2011-09-12
Reviewed by Ryosuke Niwa.

Source/WebCore:

execCommand("InsertUnorderedList") was crashing if the parent node of the target is
a kind of list element and it is not contenteditable.
This patch checks the parent node is contenteditable.

Test: editing/execCommand/insert-list-in-noneditable-list-parent.html

* editing/htmlediting.cpp:
(WebCore::enclosingListChild): Checks the parent node is contenteditable.

LayoutTests:

Added tests.

* editing/execCommand/insert-list-in-noneditable-list-parent-expected.txt: Added.
* editing/execCommand/insert-list-in-noneditable-list-parent.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@95017 268f45cc-cd09-0410-ab3c-d52691b4dbfc