Sites that use history pushState or replaceState are recorded in history in Private Browsing mode. https://bugs.webkit.org/show_bug.cgi?id=68208
Reviewed by Brady Eidson.

Source/WebCore:

Test: TestWebKitAPI/Tests/WebKit2/PrivateBrowsingPushStateNoHistoryCallback.

Add the privateBrowsingEnabled check that is used in other places in HistoryController
before calling addVisitedLink and FrameLoaderClient::updateGlobalHistory.

* loader/HistoryController.cpp:
(WebCore::HistoryController::pushState):
(WebCore::HistoryController::replaceState):

Tools:

Add a test that the WKContextDidNavigateWithNavigationDataCallback (which is invoked by
updateGlobalHistory) is not called in Private Browsing when the page does history.pushState.

* TestWebKitAPI/Tests/WebKit2/PrivateBrowsingPushStateNoHistoryCallback.cpp: Added.
(TestWebKitAPI::didNavigateWithNavigationData):
This is a context history client callback, so it should not be called when in Private
Browsing. Add a call to FAIL().
(TestWebKitAPI::didSameDocumentNavigationForFrame):
This is a page load client callback that happens after the didNavigateWithNavigationData
callback. End the test.
(TestWebKitAPI::TEST):
Enable Private Browsing and load a page that uses history.pushState.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
Add PrivateBrowsingPushStateNoHistoryCallback.cpp and push-state.html.
* TestWebKitAPI/Tests/WebKit2/push-state.html: Added.
Ditto.
* TestWebKitAPI/win/TestWebKitAPI.vcproj:
Ditto.
* TestWebKitAPI/win/copy-resources.cmd:
Ditto.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@95440 268f45cc-cd09-0410-ab3c-d52691b4dbfc