[Chromium] Fix expectations for two accessibility tests. https://bugs.webkit.org/show_bug.cgi?id=69827
These tests were never run on Chromium before, so they're not new
failures. aria-roles had different expectations on different Mac
platforms, but on Chromium the results will always be the same, so
added Chromium-specific expectations. image-map-update-parent is
timing out, the cause here is unclear but it was never running before
so there's no new bug.

Unreviewed.

* platform/chromium/accessibility/aria-roles-expected.txt: Added.
* platform/chromium/test_expectations.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@97164 268f45cc-cd09-0410-ab3c-d52691b4dbfc