REGRESSION (r95381): Standalone video can be focused and draws a focus ring. https://bugs.webkit.org/show_bug.cgi?id=69097
Patch by Deepak Sherveghar <bpwv64@motorola.com> on 2011-10-13
Reviewed by Eric Carlson.

Source/WebCore:

We don't want to focus a media element in a standalone document.

Test: fast/events/media-focus-in-standalone-media-document.html

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::supportsFocus): return false if media element is in a standalone media document.

LayoutTests:

New test to check that media element in a standalone document are not focussed.

* fast/events/media-focus-in-standalone-media-document-expected.txt: Added.
* fast/events/media-focus-in-standalone-media-document.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@97367 268f45cc-cd09-0410-ab3c-d52691b4dbfc