CSS Counters have wrong values https://bugs.webkit.org/show_bug.cgi?id=69605
Reviewed by Darin Adler.

Source/WebCore:

Test: fast/css/counters/after-continuation.html

Added a new method for getting the renderer of the "after"
pseudo-element that handles continuations.
Hooked up the new method with the CSS counter code.

* rendering/RenderCounter.cpp:
(WebCore::rendererOfAfterPseudoElement):
(WebCore::previousInPreOrder):
(WebCore::nextInPreOrder):
* rendering/RenderObject.h:

LayoutTests:

This test highlights the ability of the CSS counter handling code to
find a counter attached to a continuation of a renderer.

* fast/css/counters/after-continuation-expected.txt: Added.
* fast/css/counters/after-continuation.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@97927 268f45cc-cd09-0410-ab3c-d52691b4dbfc