[v8] Use throwError instead of compiling and running script in handleMaxRecursionDepthExceeded https://bugs.webkit.org/show_bug.cgi?id=72432
Reviewed by Adam Barth.

Source/WebCore:

* bindings/v8/V8Proxy.cpp:
(WebCore::handleMaxRecursionDepthExceeded):

LayoutTests:

Updated test that depends on the error message. The new message is a
strict improvement, as it includes the line number at which the
too-deep call was triggered. Also, we only expect a single error
message; it's not clear why the previous version generated a dup.

Also consolidate expectations, as the test does not differ per platform.

* platform/chromium-cg-mac/fast/xmlhttprequest/xmlhttprequest-recursive-sync-event-expected.txt: Removed.
* platform/chromium-mac/fast/xmlhttprequest/xmlhttprequest-recursive-sync-event-expected.txt: Removed.
* platform/chromium-win/fast/xmlhttprequest/xmlhttprequest-recursive-sync-event-expected.txt: Removed.
* platform/chromium/fast/xmlhttprequest/xmlhttprequest-recursive-sync-event-expected.txt: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@100376 268f45cc-cd09-0410-ab3c-d52691b4dbfc