[chromium] HDC leak in Uniscribe Helper https://bugs.webkit.org/show_bug.cgi?id=68598
Patch by Marc-Andre Decoste <mad@chromium.org> on 2011-11-23
Reviewed by Darin Fisher.

For some reason the Script functions on Windows sometimes return
E_PENDING even with a non-NULL DC, so we must handle that case.
Also, we should not use the screen DC to select font since this
refreshes the whole desktop, so I added a cached compatible DC.

Note that this doesn't reproduce with WebKit alone, it only reproduces
within Chrome, so we can't write a WebKit test for it. A chromium
browser test will be added once this change gets rolled in the
chromium DEPS file.

* platform/graphics/chromium/UniscribeHelper.cpp:
(WebCore::UniscribeHelper::shape):
(WebCore::UniscribeHelper::EnsureCachedDCCreated):
(WebCore::UniscribeHelper::fillShapes):
* platform/graphics/chromium/UniscribeHelper.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@101059 268f45cc-cd09-0410-ab3c-d52691b4dbfc