RenderLayer::updateZOrderLists(): Inline early-return condition. <http://webkit.org/b/74013>
Reviewed by Anders Carlsson.

updateZOrderLists() was hot (1.2%) when scrolling around on youtube.com,
and 85% of the calls were taking the early return path.

Inlined the two checks for the early return and renamed the function
to updateZOrderListsSlowCase(). Also reversed their order to avoid the
virtual call (RenderObject::isRenderView()) if possible.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateZOrderListsSlowCase):
* rendering/RenderLayer.h:
(WebCore::RenderLayer::updateZOrderLists):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@102266 268f45cc-cd09-0410-ab3c-d52691b4dbfc