[EFL][WK2] Remove codes to move and resize the clip. https://bugs.webkit.org/show_bug.cgi?id=75428
Patch by Eunmi Lee <eunmi15.lee@samsung.com> on 2012-01-10
Reviewed by Andreas Kling.

Remove codes to move and resize the clip because clip is not used now.
We have to show whole area of evas_object_image so we don't have to
clip any area.
In the WebKit1 Efl, the clip is used to clip the area to show during
weak zoom, so we can use codes to manipulate the clip when we want to
implement weak zoom for WebKit2 Efl.

* UIProcess/API/efl/ewk_view.cpp:
(_ewk_view_smart_calculate):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@104549 268f45cc-cd09-0410-ab3c-d52691b4dbfc