Filtered element with composited content beneath it must be composited https://bugs.webkit.org/show_bug.cgi?id=76322
Source/WebCore:

Reviewed by Dean Jackson.

If a RenderLayer has a filter effect, and a composited descendant, then
that layer must also be composited so that the filter is applied via
the compositing system, rather than via painting (otherwise the filter
will not affect the descendant).

Test: css3/filters/filtered-compositing-descendant.html

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::requiresCompositingWhenDescendantsAreCompositing):

LayoutTests:

Reviewed by Dean Jackson.

Testcase with a filtered element having a composited child.

* css3/filters/filtered-compositing-descendant-expected.png: Added.
* css3/filters/filtered-compositing-descendant-expected.txt: Added.
* css3/filters/filtered-compositing-descendant.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@105090 268f45cc-cd09-0410-ab3c-d52691b4dbfc