https://bugs.webkit.org/show_bug.cgi?id=76245 [Mac][DRT] should allow query string for the file URL.
Patch by Hajime Morrita <morrita@chromium.org> on 2012-01-17
Reviewed by Tony Chang.

Taught DRT about file:// as a url instead of a file path.
We are now able to use a path file:///foo/bar.html?baz as a url
with which DRT opens "/foo/bar.html" with "baz" as a query string.

* DumpRenderTree/mac/DumpRenderTree.mm:
(runTest):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@105218 268f45cc-cd09-0410-ab3c-d52691b4dbfc