CSS Shaders: Remove the setTimeout from the layout tests https://bugs.webkit.org/show_bug.cgi?id=76535
Reviewed by Tony Chang.

Source/WebCore:

We had setTimeout on old tests because the snapshot picture was taken too early, before the shaders were loaded.
The problem was that the RenderLayer was notified that the shader was loaded only after the onload event was triggered,
so a simple setTimeout(0) would have fixed the issue, but better than that would be to to always call CachedResource::data
in CachedShader::data, which notifies the load earlier (before onload).

No new tests, just removed the setTimeout from old ones.

* loader/cache/CachedShader.cpp:
(WebCore::CachedShader::data):

LayoutTests:

No need for setTimeout anymore, so I've removed them from the custom filter tests.

* css3/filters/custom-filter-shader-cache.html:
* css3/filters/effect-custom-combined-missing.html:
* css3/filters/effect-custom-parameters.html:
* css3/filters/effect-custom.html:
* css3/filters/missing-custom-filter-shader.html:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@105454 268f45cc-cd09-0410-ab3c-d52691b4dbfc