"text" and "URL" legacy clipboard types should not be case sensitive https://bugs.webkit.org/show_bug.cgi?id=76947
Reviewed by Adam Barth.

Source/WebCore:

This matches the HTML5 spec which says the first thing to do is to lowercase
the type before comparing.
I also removed the stripping of whitespace since that is not part of the modern spec (and no tests failed as a result of removal).
Turns out we already had a test for case sensitivity, but it was disabled on Mac
so I just re-enabled it and have added failing expectations for the parts
of the test we still fail.

* platform/mac/ClipboardMac.mm:
(WebCore::cocoaTypeFromHTMLClipboardType):
(WebCore::ClipboardMac::getData):

LayoutTests:

This test has been disabled on Mac since it was added 2 years ago.
But we actually need to fix at least one of the bugs covered by this
test, so I'm unskipping it and adding failing expectations for the
parts Mac doens't yet pass.

* platform/mac/Skipped:
* platform/mac/editing/pasteboard/dataTransfer-setData-getData-expected.txt: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@105941 268f45cc-cd09-0410-ab3c-d52691b4dbfc