[chromium] Remove unnecessary retry logic in LayerRendererChromium initialization for accelerated painting https://bugs.webkit.org/show_bug.cgi?id=77247
Reviewed by Kenneth Russell.

The accelerate painting setting is done differently from other capability-dependent settings for no good reason,
requiring that we retry initialization with different settings. For all other settings we set the capabilities
bit to true if the setting is requested and if the required capabilities exist on the underlying context.

* platform/graphics/chromium/LayerRendererChromium.cpp:
(WebCore::LayerRendererChromium::create):
(WebCore::LayerRendererChromium::initialize):
* platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp:
(WebCore::CCLayerTreeHostImpl::initializeLayerRenderer):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@106306 268f45cc-cd09-0410-ab3c-d52691b4dbfc