[Chromium] ChromeClientImpl::dispatchViewportPropertiesDidChange is repeatedly called in Google News https://bugs.webkit.org/show_bug.cgi?id=77429
Reviewed by Darin Fisher.

m_webView->isPageScaleFactorSet() is tested in ChromeClientImpl::layoutUpdated, and it is never true,
because WebViewImpl::setPageScaleFactorPreservingScrollOffset is called repeatedly and exits early
and thus never calls WebViewImpl::setPageScaleFactor which sets the flag.

The simplest solution is to simply get rid of the early return in
WebViewImpl::setPageScaleFactorPreservingScrollOffset.

* src/WebViewImpl.cpp:
(WebKit::WebViewImpl::setPageScaleFactorPreservingScrollOffset):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@106395 268f45cc-cd09-0410-ab3c-d52691b4dbfc