StringProtoFuncToUpperCase should call StringImpl::upper similar to StringProtoToLowerCase https://bugs.webkit.org/show_bug.cgi?id=76647
Reviewed by Darin Adler.

Changed stringProtoFuncToUpperCase to call StringImpl::upper() in a manor similar
to stringProtoFuncToLowerCase().  Fixed StringImpl::upper() to handle to special
cases.  One case is s-sharp (0xdf) which converts to "SS".  The other case is 
for characters which become 16 bit values when converted to upper case.  For
those, we up convert the the source string and use the 16 bit path.

* runtime/StringPrototype.cpp:
(JSC::stringProtoFuncToUpperCase):
* wtf/text/StringImpl.cpp:
(WTF::StringImpl::upper):
* wtf/unicode/CharacterNames.h:
(smallLetterSharpS): New constant


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@106417 268f45cc-cd09-0410-ab3c-d52691b4dbfc