[GTK] Add DRT support for modal dialogs https://bugs.webkit.org/show_bug.cgi?id=53600
Patch by Zan Dobersek <zandobersek@gmail.com> on 2012-02-09
Reviewed by Martin Robinson.

Source/WebKit/gtk:

Add support for running modal dialogs.

A new signal is added to the WebKitWebView. When emitted, the web view
is requesting to be displayed as a modal dialog. If accepted, the
ChromeClient then establishes a loop that is run until the chrome is destroyed.
If denied, nothing is done.

* WebCoreSupport/ChromeClientGtk.cpp:
(WebKit::ChromeClient::ChromeClient):
(WebKit::ChromeClient::chromeDestroyed):
(WebKit::ChromeClient::canRunModal):
(WebKit::ChromeClient::runModal):
* WebCoreSupport/ChromeClientGtk.h:
(ChromeClient):
* webkit/webkitwebview.cpp:
(webkit_web_view_class_init):

Tools:

Implement the modal dialogs handling in DumpRenderTree and GtkLauncher.
Most significantly, when creating a new web view in GtkLauncher, the window
of the web view opener is passed around. In DumpRenderTree, the window of
the main web view is always considered as the opener. This is required so
the modal dialog's toplevel can be made transient for the opener's toplevel.

* DumpRenderTree/gtk/DumpRenderTree.cpp:
(webViewRunModalDialog):
(createWebView):
* GtkLauncher/main.c:
(createWebViewCb):
(runModalDialogCb):
(createBrowser):
(createWindow):
(main):

LayoutTests:

Unskip the passing modal dialogs tests and log the remaining skipped test.

* platform/gtk/Skipped:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@107351 268f45cc-cd09-0410-ab3c-d52691b4dbfc