Allow matched property cache for elements with additional attribute style. <http://webkit.org/b/79583>
Reviewed by Antti Koivisto.

There's no reason to disallow the matched style property cache for elements
that return something from additionalAttributeStyle(). The only requirement
for a property set to be cached is that it either doesn't mutate OR that it
invalidates the document's CSSStyleSelector when doing so.

This allows some more match caching for table-related elements, though we
are still held back by explicitly 'inherited' properties in html.css.

* css/CSSStyleSelector.cpp:
(WebCore::CSSStyleSelector::collectMatchingRulesForList):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@108916 268f45cc-cd09-0410-ab3c-d52691b4dbfc