Some run-javascriptcore-tests broken for 32-bit debug https://bugs.webkit.org/show_bug.cgi?id=79844
Rubber stamped by Oliver Hunt.
        
These assertions are just plain wrong for 32-bit. We could either have a massive
assertion that depends on value representation, that has to be changed every
time we change the JITs, resulting in a bug tail of debug-mode crashes, or we
could get rid of the assertions. I pick the latter.

* dfg/DFGOperations.cpp:
* jit/JITStubs.cpp:
(JSC::DEFINE_STUB_FUNCTION):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@109172 268f45cc-cd09-0410-ab3c-d52691b4dbfc