Applying region style should not need to access parent rules https://bugs.webkit.org/show_bug.cgi?id=79910
Reviewed by Andreas Kling.

Currently CSSStyleSelector::applyProperties looks into parent rules to see if a rule is
part of region style. The plan is to eliminate the rule parent pointer so this needs to be refactored.
        
Add a bit to RuleData to indicate if the StyleRule is part of a region style.

* css/CSSStyleSelector.cpp:
(RuleData):
(WebCore::RuleData::isInRegionRule):
(RuleSet):
(WebCore::CSSStyleSelector::addMatchedProperties):
(WebCore::CSSStyleSelector::sortAndTransferMatchedRules):
(WebCore::CSSStyleSelector::collectMatchingRulesForList):
* css/CSSStyleSelector.h:
(CSSStyleSelector):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@109227 268f45cc-cd09-0410-ab3c-d52691b4dbfc