chromium: start running virtual gpu tests https://bugs.webkit.org/show_bug.cgi?id=80050
Unreviewed, expectations change.

This patch un-SKIPs the platform/chromium/virtual/gpu tests
and copies their expectations from the GPU and generic
expectations for fast/canvas and canvas/philip. I
will re-sort the expectations in subsequent patches so
that the virtual expectations are next to the action ones,
for easier maintenance.

* platform/chromium/test_expectations.txt:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@109447 268f45cc-cd09-0410-ab3c-d52691b4dbfc