Merge expectations for svg/as-background-image/svg-as-background-6.html.  Previously, Snow Leopard had different expectations than every other platform, but the behavior seems to match the expectations for other platforms.
Really what's going on is that this is a bad test for the reasons
explained in the test_expectations.txt file.  I'm not entirely sure how
to fix the test to have repeatable results.

This patch also notes some more compositing reflection tests as flaky,
similar to the failures in
https://bugs.webkit.org/show_bug.cgi?id=79823.  These tests have been
flaky for a while, but they seem to be failing more frequently now.  I
couldn't find any recent changes to explain why that might be.

* platform/chromium/test_expectations.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@110621 268f45cc-cd09-0410-ab3c-d52691b4dbfc