Heap::destroy leaks CopiedSpace https://bugs.webkit.org/show_bug.cgi?id=81055
Reviewed by Geoffrey Garen.

Added a destroy() function to CopiedSpace that moves all normal size 
CopiedBlocks from the CopiedSpace to the Heap's list of free blocks 
as well as deallocates all of the oversize blocks in the CopiedSpace. 
This function is now called in Heap::destroy().

* heap/CopiedSpace.cpp:
(JSC::CopiedSpace::destroy):
(JSC):
* heap/CopiedSpace.h:
(CopiedSpace):
* heap/Heap.cpp:
(JSC::Heap::destroy):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@110748 268f45cc-cd09-0410-ab3c-d52691b4dbfc