Various autoresize fixes. https://bugs.webkit.org/show_bug.cgi?id=81199
Reviewed by Darin Fisher.

Source/WebCore:

Tested by WebViewTest additions.

* page/FrameView.cpp:
(WebCore::FrameView::autoSizeIfEnabled): Fixed a bug that caused a horizontal
scrollbar to appear when the size was width was fixed and height overflowed which
is one of the tests that was added.

Source/WebKit/chromium:

* public/WebView.h:
(WebView): Remove a deprecated api.
* src/WebViewImpl.cpp: Ditto.
* src/WebViewImpl.h: Ditto.
* tests/FrameTestHelpers.cpp:
(WebKit::FrameTestHelpers::registerMockedURLLoad): Added an overload with more flexibility.
The previous api didn't allow one to request a url with a search string.
(FrameTestHelpers):
* tests/FrameTestHelpers.h:
(FrameTestHelpers): Ditto.
* tests/WebViewTest.cpp: Fix the test because calling javascript:
was working. Make the test generic and added a lot more tests to cover
various possibilities. I avoided using a layout test because there wasn't
a way to get the scrollbar state, so I just generalized a test that was already there.
* tests/data/specify_size.html: Allow the test to take its size from the search string.

Tools:

* DumpRenderTree/chromium/LayoutTestController.cpp:
(LayoutTestController::reset): Stop using the deprecated api that was removed.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@110875 268f45cc-cd09-0410-ab3c-d52691b4dbfc