REGRESSION(r110072): Improper repainting on table with overflow: hidden https://bugs.webkit.org/show_bug.cgi?id=81460
Reviewed by Simon Fraser.

Source/WebCore:

Test: fast/repaint/table-hover-on-link.html

* rendering/RenderTable.cpp:
(WebCore::RenderTable::layout): Add a call to RenderBlock::updateCachedSizeForOverflowClip
so that we properly keep our cached size up to date. This fixes the problem for now but the
real solution is to make sure it is called on all classes that needs to. This is tracked by
bug 81461.

LayoutTests:

* fast/repaint/table-hover-on-link-expected.txt: Added.
* fast/repaint/table-hover-on-link.html: Added.

* platform/chromium-linux/fast/repaint/table-hover-on-link-expected.png: Added.
* platform/chromium-mac/fast/repaint/table-hover-on-link-expected.png: Added.
* platform/mac/fast/repaint/table-hover-on-link-expected.png: Added.
There seem to be some small color difference here (maybe color profile related).


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@111360 268f45cc-cd09-0410-ab3c-d52691b4dbfc