Inline StringImpl::find(UChar, ...) https://bugs.webkit.org/show_bug.cgi?id=83737
Patch by Benjamin Poulain <bpoulain@apple.com> on 2012-04-12
Reviewed by Geoffrey Garen.

The implementation of StringImpl::find() is a simple branch before invoking one of
two inline functions. The overhead of having a function for StringImpl::find() is significant.

It is better to let the compiler decide if that should be inlined or not.

* wtf/text/StringImpl.cpp:
* wtf/text/StringImpl.h:
(StringImpl):
(WTF::find):
(WTF):
(WTF::reverseFind):
(WTF::StringImpl::find):
* wtf/text/WTFString.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@114071 268f45cc-cd09-0410-ab3c-d52691b4dbfc