[chromium] Consistent checking for clipped rects when we need the computed result enclosed within the real result https://bugs.webkit.org/show_bug.cgi?id=83543
Reviewed by Adrienne Walker.

It should not be possible to make a rect in layer space that is clipped
by the camera but for which the screen space transform gives a
rectilinear output. But use consistent methods for checking that the
result remains enclosed within the actual pixels.

One day when clipped is true, we can find an interior axis-aligned rect
within the clipped result, and checking clipped explicitly makes this
more clear.

Covered by existing tests.

* platform/graphics/chromium/cc/CCOcclusionTracker.cpp:
(WebCore::computeOcclusionBehindLayer):
(WebCore::::markOccludedBehindLayer):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@114383 268f45cc-cd09-0410-ab3c-d52691b4dbfc