Crash in -webkit-calc https://bugs.webkit.org/show_bug.cgi?id=86954
Reviewed by Abhishek Arya.

Source/WebCore:

Ensure that both ends of a blend are checked for being Calculated. The current
check did not cover the case where the 'from' end was Calculated and nonzero,
and the 'to' end was zero and not Calculated.

Blending doesn't work with calculated expressions and is being tracked in
bug http://webkit.org/b/86160

Tests: css3/calc/transition-crash3.html
       css3/calc/transition-crash4.html

* platform/Length.h:
(WebCore::Length::blend):

LayoutTests:

* css3/calc/transition-crash3-expected.txt: Added.
* css3/calc/transition-crash3.html: Added.
* css3/calc/transition-crash4-expected.txt: Added.
* css3/calc/transition-crash4.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@117748 268f45cc-cd09-0410-ab3c-d52691b4dbfc