Body scrollWidth() and scrollHeight() should be page scale-invariant https://bugs.webkit.org/show_bug.cgi?id=87494
RenderView::documentRect() is calculating the "scaled" document rect by applying
the current transformation matrix to the unscaledDocumentRect() and then
returning the rounded-out IntRect result.

This rounding out is incorrect because it allows the scaled rectangle to
represent an area that is not actually covered by the document.

We fix this by applying the current transform to the document rect
as a FloatRect and then explicitly converting to IntRect, which
takes the floor of the resulting rectangle coordinates instead of
rounding them out.

This is evidenced by the document.body.scrollWidth() and
document.body.scrollHeight() changing under page scale factor when
they are expected to remain invariant.

Reviewed by James Robinson.

Source/WebCore:

Test: fast/dom/window-scroll-scaling.html

* rendering/RenderView.cpp:
(WebCore::RenderView::documentRect):

LayoutTests:

* fast/dom/window-scroll-scaling-expected.txt: Added.
* fast/dom/window-scroll-scaling.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@118569 268f45cc-cd09-0410-ab3c-d52691b4dbfc