[Qt] Sync up favicon-implementation with WebView url changes in r118158
https://bugs.webkit.org/show_bug.cgi?id=87133

We now base64-encode the page url in the image-provider url, so that any
normalization done by QUrl will not mess up the page-url. The logic of
creating and parsing the provider-url has been moved into the image
provider, to keep it in one place.

We were also releasing icons (even ones we hadn't retained), which we can't
do since we don't know when the icon url is no longer in use.

Reviewed-by Simon Hausmann.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@118762 268f45cc-cd09-0410-ab3c-d52691b4dbfc