[V8] Clean up V8LazyEventListener to use one less function call https://bugs.webkit.org/show_bug.cgi?id=87785
Reviewed by Kentaro Hara.

Instead of relying on 'arguments' we rely on 'this'. 'this' cannot be intercepted by a
with-statement so it is safe to use that to bind the context objects. This allows us to
remove one layer of function indirection in the generated code.

Source/WebCore:

Covered by existing tests.

* bindings/v8/V8LazyEventListener.cpp:
(WebCore::V8LazyEventListener::prepareListenerObject):

LayoutTests:

* platform/chromium-win/inspector/debugger/debugger-scripts-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@118944 268f45cc-cd09-0410-ab3c-d52691b4dbfc