InsertUnorderedList and InsertOrderedList (execCommand) do not remove bullets https://bugs.webkit.org/show_bug.cgi?id=84597
Patch by Arpita Bahuguna <arpitabahuguna@gmail.com> on 2012-06-07
Reviewed by Ryosuke Niwa.

Source/WebCore:

Comparison between start and end VisiblePositions will fail when a position inside
a paragraph is compared with one at the start of the paragraph.

This fix thus converts the end VisiblePosition (which could be inside the paragraph)
to the start of the paragraph thereby enabling a proper comparison of start and end
positions.

Test: editing/execCommand/remove-list-from-multi-list-items.html

* editing/InsertListCommand.cpp:
(WebCore::InsertListCommand::selectionHasListOfType):

LayoutTests:

This test verifies removing of an (un)ordered list on a selection consisting of multiple
list items (paragraphs) when the end point of the selection points to the end of the last
list item; followed by another paragraph not part of the selection/list.

* editing/execCommand/remove-list-from-multi-list-items-expected.txt: Added.
* editing/execCommand/remove-list-from-multi-list-items.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@119701 268f45cc-cd09-0410-ab3c-d52691b4dbfc