LayoutTests/http/tests/xmlhttprequest/origin-exact-matching.html is giving trouble due to too many redirects https://bugs.webkit.org/show_bug.cgi?id=88896
Patch by Pablo Flouret <pablof@motorola.com> on 2012-06-12
Reviewed by Darin Adler.

The test was giving out "too many redirect" errors on GTK and EFL due to
using redirect-cors.php without an actual url to redirect.

Redirects are not actually tested in that test, a script that set the
Access-Control-Allow-Origin header to an arbitrary origin was needed and
redirect-cors.php was the only existing one that had the capability. The
infinite redirect thing was an unintended consequence.
Use resources/access-control-allow-lists.php to set the header instead.

* http/tests/xmlhttprequest/origin-exact-matching-expected.txt:
* http/tests/xmlhttprequest/resources/origin-exact-matching-iframe.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@120107 268f45cc-cd09-0410-ab3c-d52691b4dbfc