Fix link error with fillWithEmptyClients https://bugs.webkit.org/show_bug.cgi?id=89484
Patch by Alexandre Elias <aelias@google.com> on 2012-06-19
Reviewed by Ryosuke Niwa.

fillWithEmptyClients() was within #if ENABLE() guard in the .cpp file
but not in the header.  There is now a caller unrelated to these ENABLEs
in WebHelperPluginImpl, so linking fails if both SVG and PAGE_POPUP
are disabled.

This function looks generic, so just remove the ENABLE guards.

No new tests. (Build fix.)

* loader/EmptyClients.cpp:
(WebCore::fillWithEmptyClients):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@120741 268f45cc-cd09-0410-ab3c-d52691b4dbfc