[Mac][WK2] Enable HTTPS tests https://bugs.webkit.org/show_bug.cgi?id=90600
Reviewed by Brady Eidson.

Tools:

* DumpRenderTree/mac/DumpRenderTree.mm: (dumpRenderTree): Removed a misleading comment.
It explained why we were doing this here instead of a frame load delegate, but that's
just history, not an explanation of why we need this code.

* WebKitTestRunner/InjectedBundle/mac/InjectedBundleMac.mm: (WTR::InjectedBundle::platformInitialize):
Ignore certificate errors for localhost and 127.0.0.1. Since this is only needed
for tests, it seems appropriate to have platform specific solutions in every WTR,
and not add a WebKit2 API.

LayoutTests:

Unskip in common wk2 skipped file, re-skip everywhere besides Mac.

* platform/gtk-wk2/Skipped:
* platform/qt-5.0-wk2/Skipped:
* platform/win-wk2/Skipped:
* platform/wk2/Skipped:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@121913 268f45cc-cd09-0410-ab3c-d52691b4dbfc