TOUCH_ADJUSTMENT is too aggressive when snapping to large elements. https://bugs.webkit.org/show_bug.cgi?id=91262
Patch by Kevin Ellis <kevers@chromium.org> on 2012-07-18
Reviewed by Antonio Gomes.

Source/WebCore:

Constrains the extent to which the touch point can be adjusted when
generating synthetic mouse events when TOUCH_ADJUSTEMNT is enabled.
Previously, the target position snapped to the center of the target
element, which can be far removed from the touch position when tapping
on or near a large element.  The refined strategy is to leave the
adjusted position unchanged if tapping within the element or to snap
to the center of the overlap region if the touch point lies outside the
bounds of the element, but the touch area and element bounds overlap.
For non-rectilineary bounds, a point lying outside the element boundary
is pulled towards the center of the element, by an amount limited by
the radius of the touch area.

Tests: touchadjustment/big-div.html
       touchadjustment/rotated-node.html

* page/TouchAdjustment.cpp:
(WebCore::TouchAdjustment::contentsToWindow):
(TouchAdjustment):
(WebCore::TouchAdjustment::snapTo):
(WebCore::TouchAdjustment::findNodeWithLowestDistanceMetric):

LayoutTests:

Adding a test case to ensure that the adjusted touch position is
within the bounds of the target element and the touch area.
Previously, the target position snapped to the center of the target
element, which can be far removed from the touch area.

The second test is for non-rectilinear elements, and verifies that
the touch area must overlap the true bounds of the element for an
adjustment to occur.

* touchadjustment/big-div-expected.txt: Added.
* touchadjustment/big-div.html: Added.
* touchadjustment/rotated-node-expected.txt: Added.
* touchadjustment/rotated-node.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@122970 268f45cc-cd09-0410-ab3c-d52691b4dbfc