[CSSRegions]Crash when flowing a region into itself https://bugs.webkit.org/show_bug.cgi?id=90289
Reviewed by Andreas Kling.

If a region is designed to receive itself as content through a named flow, we avoid
a circular dependency by not attaching the region to the desired named flow.
In such occasions, the region's attached flow thread object is null and
we have to check this pointer before using it in RenderRegion::styleDidChange.

Source/WebCore:

Test: fast/regions/flow-region-into-itself-crash.html

* rendering/RenderRegion.cpp:
(WebCore::RenderRegion::styleDidChange):

LayoutTests:

* fast/regions/flow-region-into-itself-crash-expected.txt: Added.
* fast/regions/flow-region-into-itself-crash.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@123196 268f45cc-cd09-0410-ab3c-d52691b4dbfc