[GTK] Default signal handler for WebKitWebView::should-show-delete-interface-for-element overrides default result https://bugs.webkit.org/show_bug.cgi?id=93600
Patch by Martin Robinson <mrobinson@igalia.com> on 2012-08-13
Reviewed by Xan Lopez.

Instead of using the default editing signal handler for ::should-show-delete-interface-for-element,
do not use a default signal handler. This means that the result of the signal defaults to FALSE,
which is the expected value to ensure that the delete interface is not shown.

* webkit/webkitwebview.cpp:
(webkit_web_view_class_init): Do not install a default signal handler.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@125407 268f45cc-cd09-0410-ab3c-d52691b4dbfc