Unreviewed GTK gardening.
Adding image failure expectation for fast/overflow/line-clamp-and-columns.html,
the reftest is failing since it was introduced in r125635.

Skipping media/video-controls-fullscreen-volume.html for the moment after
it's been failing since introduced in r125590, the failure is probably expected
and will be marked as WONTFIX when confirmed.

Marking fast/js/random-array-gc-stress.html as a flaky crasher. The crashing
occurs due to parallel GC being enabled.

* platform/gtk/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@125657 268f45cc-cd09-0410-ab3c-d52691b4dbfc