When kerning is enabled, word spacing is doubly accounted for in RenderText::computePreferredLogicalWidths https://bugs.webkit.org/show_bug.cgi?id=97146
Reviewed by Anders Carlsson.

Source/WebCore: 

Test: fast/text/word-space-with-kerning.html

* rendering/RenderText.cpp:
(WebCore::RenderText::computePreferredLogicalWidths): When kerning is enabled, words are
measured with their trailing space, if there is one, then the width of a space is subtracted.
Changed that width, stored in the local variable wordTrailingSpaceWidth, to include the
word spacing, since it is included in the result of widthFromCache().

LayoutTests: 

* fast/text/word-space-with-kerning-expected.html: Added.
* fast/text/word-space-with-kerning.html: Added.



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@129068 268f45cc-cd09-0410-ab3c-d52691b4dbfc