[Qt] [WK1] Spaces missing in output of fast/forms/mailto/advanced-{get,put}.html https://bugs.webkit.org/show_bug.cgi?id=92551
Patch by Marcelo Lira <marcelo.lira@openbossa.org> on 2012-09-20
Reviewed by Adam Barth.

The problem here is that the text input boxes are rendered differently
in chromium and qt/gtk/efl, causing the text dump to also differ.

There are 4 adjacent input boxes in a page 800 pixel wide, the width
of these inputs is unspecified, as it should since this is unrelated
to purpose of the test, and left for the port to decide. Chromium puts
the four in the same line, qt/gtk/efl breaks the line before the fourth
input, causing an extra space to appear in the text dump.

The test was corrected by defining a fixed width for the input
elements, ensuring the text dumps will be identical in all platforms,
eliminating the need for special expected files.

* fast/forms/mailto/advanced-get.html:
* fast/forms/mailto/advanced-put.html:
* platform/efl/fast/forms/mailto/advanced-get-expected.txt: Removed.
* platform/efl/fast/forms/mailto/advanced-put-expected.txt: Removed.
* platform/gtk/fast/forms/mailto/advanced-get-expected.txt: Removed.
* platform/gtk/fast/forms/mailto/advanced-put-expected.txt: Removed.
* platform/qt-5.0-wk1/Skipped:
* platform/qt-5.0-wk2/fast/forms/mailto/advanced-get-expected.txt: Removed.
* platform/qt-5.0-wk2/fast/forms/mailto/advanced-put-expected.txt: Removed.
* platform/qt-mac/Skipped:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@129120 268f45cc-cd09-0410-ab3c-d52691b4dbfc