We should not expect percent-escaped hostname for invalid URLs https://bugs.webkit.org/show_bug.cgi?id=97415
Reviewed by Adam Barth.

The URL layout tests encode some Chromium specific behavior.
Chromium does a partial canonicalization of invalid URLs. Neither Safari, nor Firefox nor Opera
do that.

This patch fix the tests and expectations for the IP tests.

* fast/url/invalid-urls-utf8-expected.txt:
* fast/url/invalid-urls-utf8.html:
Move the "empty authority" check from both IP tests to the invalid URLs test. An empty authority
for a standard URL causes an invalid URL, it should be kept as is.

* fast/url/ipv4-expected.txt:
* fast/url/ipv6-expected.txt:
* fast/url/script-tests/ipv4.js:
* fast/url/script-tests/ipv6.js: Extend a comment to clarify the rule.
Invalid URLs should not be escaped.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@129445 268f45cc-cd09-0410-ab3c-d52691b4dbfc