[GTK] Make inspector directly useable in GtkLauncher/MiniBrowser https://bugs.webkit.org/show_bug.cgi?id=98310
Patch by Adrian Perez de Castro <aperez@igalia.com> on 2012-10-03
Reviewed by Martin Robinson.

Make MiniBrowser/GtkLauncher define the path to the inspector
resources by setting the WEBKIT_INSPECTOR_PATH environment
variable pointing to the copy of the inspector resources in
the build directory. If the environment variable is already
defined, its value is not overwritten. The path is derived
from the WEBKIT_EXEC_PATH macro defined in the makefiles.

* GNUmakefile.am:
* GtkLauncher/main.c:
(main):
* MiniBrowser/gtk/main.c:
(main):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@130349 268f45cc-cd09-0410-ab3c-d52691b4dbfc