Don't duplicated build numbers when merging flakiness dashboard JSON https://bugs.webkit.org/show_bug.cgi?id=98812
Reviewed by Dirk Pranke.

In general, build numbers are not unique, but we should basically never
get the same build number uploaded twice in a row. This is a workaround
for https://bugs.webkit.org/show_bug.cgi?id=97643, which frequently hits this
because we update results_small.json and timeout updating results.json and then
we retry the whole request.

* TestResultServer/model/jsonresults.py:
(JsonResults.merge):
* TestResultServer/model/jsonresults_unittest.py:
(JsonResultsTest.test_merge_duplicate_build_number):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@130790 268f45cc-cd09-0410-ab3c-d52691b4dbfc