margin-top/bottom has no effect for child nodes of flex items https://bugs.webkit.org/show_bug.cgi?id=99923
Reviewed by Ojan Vafai.

Source/WebCore:

Flexitems, like table cells, shouldn't collapse margins.

Test: css3/flexbox/flexitem-no-margin-collapsing.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::MarginInfo::MarginInfo): Check to see if the parent is a flexible box.
We should always have a parent if we make it this far in the check.

LayoutTests:

Add a test to make sure we don't collapse margins in flex items.

* css3/flexbox/flexitem-no-margin-collapsing-expected.txt: Added.
* css3/flexbox/flexitem-no-margin-collapsing.html: Added.
* css3/flexbox/repaint.html:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@132164 268f45cc-cd09-0410-ab3c-d52691b4dbfc