Handle ArgumentCoder template specializations that take the ArgumentEncoder as a reference https://bugs.webkit.org/show_bug.cgi?id=100056
Reviewed by Andreas Kling.

Use template magic to make it possible to have ArgumentCoder specializations where the encode
function takes the ArgumentEncoder object as a reference instead of as a pointer. Also, add an
operator<< to ArgumentEncoder and change the string related ArgumentCoder specializations over to taking
the encoder as a reference and using stream operators.

* Platform/CoreIPC/ArgumentCoders.cpp:
(CoreIPC::::encode):
* Platform/CoreIPC/ArgumentCoders.h:
* Platform/CoreIPC/ArgumentEncoder.h:
(ArgumentEncoder):
(UsesDeprecatedEncodeFunction):
(NoType):
(CoreIPC::ArgumentEncoder::encode):
(CoreIPC::ArgumentEncoder::operator<<):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@132169 268f45cc-cd09-0410-ab3c-d52691b4dbfc