BasicShapePolygon::path takes width instead of height for boundary calculation https://bugs.webkit.org/show_bug.cgi?id=99919
Reviewed by Darin Adler.

Source/WebCore:

The 'y' parameters of polygon were calculated by the with of the bounding box of the object.
This caused problems on percentage values for point positions.
Changed it to the height of the bounding box.

Test: css3/masking/clip-path-polygon-percentage.html

* rendering/style/BasicShapes.cpp:
(WebCore::BasicShapePolygon::path):

LayoutTests:

Test that the correct relation is used for calculating point position of an polygon.

* css3/masking/clip-path-polygon-percentage-expected.html: Added.
* css3/masking/clip-path-polygon-percentage.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@132257 268f45cc-cd09-0410-ab3c-d52691b4dbfc