[EFL][WK2] REGRESSION(r132342): fast/events/frame-tab-focus.html fails https://bugs.webkit.org/show_bug.cgi?id=100646
Reviewed by Kenneth Rohde Christiansen.

Tools: 

* WebKitTestRunner/efl/PlatformWebViewEfl.cpp:
(WTR::PlatformWebView::focus): Improve the check we perform to
decide whether the focus should be taken out of the view before
focusing it again so that we do not unnecessarily unfocus it and
cause blur/focus events to be triggered. Instead of just verifying
if the Evas_Object representing the view has focus, use the WK C
API to check that the currently focused frame is not the main
frame.

LayoutTests: 

* platform/efl-wk2/TestExpectations: Unskip
fast/events/frame-tab-focus.html.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@132796 268f45cc-cd09-0410-ab3c-d52691b4dbfc