[Qt][WK2] Fit-to-width broken on pages with viewport meta tag https://bugs.webkit.org/show_bug.cgi?id=99715
Reviewed by Jocelyn Turcotte.

Source/WebCore:

Add a bool that indicates if the content had an explicit
initial-scale in the viewport meta tag.

* dom/ViewportArguments.cpp:
(WebCore::computeViewportAttributes):
* dom/ViewportArguments.h:
(ViewportAttributes):

Source/WebKit2:

The initial scale from the viewport attributes should only
be applied if the scale was explicitly specified in the
viewport meta tag.
If the initial scale is auto it should be calculated using
the final contents size, which might be larger than the
layout size, so that the content fits horizontally into
the view.

Also add QML unit tests that cover this functionality.

* UIProcess/API/qt/tests/qmltests/WebView/tst_fitToView.qml:
* UIProcess/API/qt/tests/qmltests/common/test5.html: Added.
* UIProcess/PageViewportController.cpp:
(WebKit::PageViewportController::PageViewportController):
(WebKit::PageViewportController::pageTransitionViewportReady):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@133624 268f45cc-cd09-0410-ab3c-d52691b4dbfc