GraphicsSurfaceGLX needs to query the drawable for YInversion. https://bugs.webkit.org/show_bug.cgi?id=101472
After r133049 the WebGL content is now displayed upside down
for the Qt port. We cannot just apply a generic rule, that
the content needs to be flipped upside down.
Instead we have to query GLX to tell us, if the drawable's
framebuffer is y-inverted.

Reviewed by Kenneth Rohde Christiansen.

* platform/graphics/surfaces/qt/GraphicsSurfaceGLX.cpp:
(WebCore::GraphicsSurfacePrivate::textureIsYInverted):
(GraphicsSurfacePrivate):
(WebCore::GraphicsSurface::platformPaintToTextureMapper):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@133874 268f45cc-cd09-0410-ab3c-d52691b4dbfc