[Chromium] fix FPS counter disappearing when changing page https://bugs.webkit.org/show_bug.cgi?id=102532
Patch by Eberhard Graether <egraether@google.com> on 2012-11-20
Reviewed by Pavel Feldman.

When changing the page with opened WebInspector and active FPS counter the WebInspector is
reloaded before the WebView switches to accelerated compositing, so the showFPSCounter call
to WebLayerTreeView is not yet possible. This fix saves the showFPSCounter value in the
WebSettings so it can be read when accelerated compositing is activated.

* src/WebViewImpl.cpp:
(WebKit::WebViewImpl::setShowFPSCounter):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@135300 268f45cc-cd09-0410-ab3c-d52691b4dbfc