[chromium] Mark last input event for current vsync interval https://bugs.webkit.org/show_bug.cgi?id=103647
Patch by Sami Kyostila <skyostil@chromium.org> on 2012-12-03
Reviewed by James Robinson.

On Android all input events are triggered at the start of a vsync to maximize
the amount of time an application has for painting. Add a field to
WebInputEvent to indicate when the input event is the final one for the
current vsync interval. This information will be used for more efficient
render scheduling.

* public/WebInputEvent.h:
(WebKit::WebInputEvent::WebInputEvent):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@136383 268f45cc-cd09-0410-ab3c-d52691b4dbfc