https://bugs.webkit.org/show_bug.cgi?id=104725 Crash in EventHandler::selectCursor() on Mac bots running fast/events/mouse-cursor-image-set.html
Reviewed by Tim Horton.

https://bugs.webkit.org/show_bug.cgi?id=99493 enabled image-set for 
the CSS cursor property. Most of the change is guarded behind #if 
ENABLE(MOUSE_CURSOR_SCALE). However, image-set was still allowed to 
parse for the cursor property as long as image-set is enabled 
(which is the current state on the Mac). This would lead to an 
inconsistent state where style information would reflect a world 
where mouse cursor scaling is enabled even though nothing else on 
the Mac would. 

The parsing should also be guarded by the cursor scaling flag.
* css/CSSParser.cpp:
(WebCore::CSSParser::parseValue):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@137367 268f45cc-cd09-0410-ab3c-d52691b4dbfc