Web Inspector: need to visually distinguish UA shadow roots https://bugs.webkit.org/show_bug.cgi?id=104877
Reviewed by Yury Semikhatsky.

Source/WebCore:

Now we have an option to show ShadowRoot in the Inspector. Since the Inspector has displayed UserAgent ShadowRoot and
Author ShadowRoot as the same #shadow-root, a user could not distinguish them.

We would like to show UserAgent ShadowRoot as #webkit-shadow-root.

Test: inspector/shadow-root.html

* inspector/InspectorDOMAgent.cpp:
(WebCore::InspectorDOMAgent::buildObjectForNode): Make nodeName of UserAgent ShadowRoot #webkit-shadow-root.
Since the correct nodeName of ShadowRoot is #document-fragment, we have to specify the nodeName of Author ShadowRoot
as #shadow-root here.

LayoutTests:

* inspector/elements/shadow-dom-modify-chardata-expected.txt:
* inspector/elements/shadow-root-expected.txt:
* inspector/elements/shadow-root.html: Adds <input> test.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@137979 268f45cc-cd09-0410-ab3c-d52691b4dbfc