Chromium: Error in AccessibilityUIElement::intValueGetterCallback https://bugs.webkit.org/show_bug.cgi?id=106682
Reviewed by Chris Fleizach.

Tools:

Modify Chromium's DRT implementation of intValue to return a
different value depending on the role.

There are actually platform-specific differences in what should
be returned in the "value" of an object, so Chromium normally keeps
these separate (i.e. valueForRange, headingLevel, hierarchicalLevel, etc.)
but this is a fine simplification to make cross-platform tests easier.

* DumpRenderTree/chromium/TestRunner/src/AccessibilityUIElementChromium.cpp:
(WebTestRunner::AccessibilityUIElement::intValueGetterCallback):

LayoutTests:

Unskip accessibility/heading-level.html now that the bug is fixed.

* platform/chromium/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@139658 268f45cc-cd09-0410-ab3c-d52691b4dbfc