[Compositor] Do not disable overlap testing for layers in front of 3D transformed layers https://bugs.webkit.org/show_bug.cgi?id=106721
Reviewed by Simon Fraser.

Source/WebCore:

RenderLayerCompositor is checking for overlapping layers and tries to figure out all the layers that need
to be drawn using composited layers. It doesn't know anything about animations yet and it only works with "start"
and "end" states. However, it knows that animations might happen, so in that cases it will avoid to do any optimizations,
as layers might get moved.

The overlapping layers optimization was also disabled for 3D layers. However, 3D layers can be computed in WebKit and
can be checked for overlap using just the "start"/"stop" states. In this patch I enable that optimization for 3D layers, too.
The result is that more layers are drawn in software mode now, so I've updated a couple of test results.

Tests: compositing/layer-creation/overlap-transformed-3d.html
       compositing/layer-creation/overlap-transformed-preserved-3d.html

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::computeCompositingRequirements):

LayoutTests:

Added two more tests for the case when 3D layers covered by a 2D layer in front. One is using flat style and the
other is using the preserve-3d mode.

Also, updated a couple of existing test results as some of the composited layers are not needed anymore.

* compositing/geometry/ancestor-overflow-change-expected.txt:
* compositing/geometry/clip-expected.txt:
* compositing/geometry/clip-inside-expected.txt:
* compositing/geometry/foreground-layer-expected.txt:
* compositing/layer-creation/overlap-animation-clipping-expected.txt:
* compositing/layer-creation/overlap-animation-container-expected.txt:
* compositing/layer-creation/overlap-transformed-3d-expected.txt: Added.
* compositing/layer-creation/overlap-transformed-3d.html: Added.
* compositing/layer-creation/overlap-transformed-preserved-3d-expected.txt: Added.
* compositing/layer-creation/overlap-transformed-preserved-3d.html: Added.
* compositing/layer-creation/overlap-transforms-expected.txt:
* compositing/overflow/clip-descendents-expected.txt:
* css3/filters/filtered-compositing-descendant-expected.txt:
* platform/chromium-win/compositing/geometry/ancestor-overflow-change-expected.txt: Removed. There should be no difference now.
* platform/chromium-win/css3/filters/filtered-compositing-descendant-expected.txt: Removed. There should be no difference now.
* platform/chromium/compositing/geometry/ancestor-overflow-change-expected.txt:
* platform/chromium/compositing/geometry/clip-expected.txt:
* platform/chromium/compositing/geometry/clip-inside-expected.txt:
* platform/chromium/compositing/geometry/foreground-layer-expected.txt:
* platform/chromium/compositing/layer-creation/overlap-animation-clipping-expected.txt:
* platform/chromium/compositing/layer-creation/overlap-animation-container-expected.txt:
* platform/chromium/compositing/layer-creation/overlap-transformed-preserved-3d-expected.txt: Added.
* platform/chromium/compositing/layer-creation/overlap-transforms-expected.txt:
* platform/chromium/compositing/overflow/clip-descendents-expected.txt:
* platform/chromium/css3/filters/filtered-compositing-descendant-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@139794 268f45cc-cd09-0410-ab3c-d52691b4dbfc