Do not track user interaction for plugins that are not snapshotted https://bugs.webkit.org/show_bug.cgi?id=107211 <rdar://problem/12967277>
Reviewed by Dean Jackson.

When the user interacts with a snapshotted plug-in, we extend that plug-in origin's
expiration date. However, plug-ins may not be snapshotted for other reasons than the
user's initial opt-in. In this case, that plug-in's origin might be not be in the cached
table. This is an edge case that was not initially considered.

* WebProcess/WebProcess.cpp:
(WebKit::WebProcess::plugInDidReceiveUserInteraction): Convert the assertion to an if
check, and return early if the entry is not found in the cache table.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@140081 268f45cc-cd09-0410-ab3c-d52691b4dbfc