[V8] Remove custom V8ScriptProfileCustom::toV8() https://bugs.webkit.org/show_bug.cgi?id=107246
Reviewed by Adam Barth.

V8ScriptProfileCustom::toV8() needs not to be custom.
Furthermore, the current custom toV8() is wrong in that
it doesn't use a creationContext and it doesn't set a
wrapper class id. This is one of steps to avoiding
ASSERT()s that will be added in
https://bugs.webkit.org/show_bug.cgi?id=107137 .

No tests. No change in behavior.

* UseV8.cmake:
* WebCore.gypi:
* bindings/v8/custom/V8ScriptProfileCustom.cpp: Removed.
* inspector/ScriptProfile.idl:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@140295 268f45cc-cd09-0410-ab3c-d52691b4dbfc