[Autotools] Remove the Canvas Path configuration option https://bugs.webkit.org/show_bug.cgi?id=107537
Reviewed by Martin Robinson.

.: 

* configure.ac: Remove the configuration option for the Canvas Path
feature as this is not in line with the new guidelines about feature
enablement in the autotools build system.

Source/WebCore: 

No new tests - no new testable functionality.

* GNUmakefile.am: Remove the Automake conditional analysis, the related
configuration option is being removed as well.
* GNUmakefile.features.am.in: Add the ENABLE_CANVAS_PATH feature define,
defaulting to 0 in release builds (but is disabled in development builds
of the GTK port as well at the moment).


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@140431 268f45cc-cd09-0410-ab3c-d52691b4dbfc