[GStreamer] Add Opus MIME type to the list of supported ones https://bugs.webkit.org/show_bug.cgi?id=98759
Patch by Adrian Perez de Castro <aperez@igalia.com> on 2013-01-23
Reviewed by Philippe Normand.

Source/WebCore:

GStreamer has support for the Opus codec (http://opus-codec.org) by
using a plugin present in the "gst-plugins-bad" repository. Opus
streams may be embedded in a Ogg container, or standalone. The case
of the Ogg container is already covered by the "*/ogg" MIME types
declared as supported by the GStreamer media playing code, but for
standalone streams to work, "audio/opus" has to be added.

With this, and the needed GStreamer plugin installed, the GTK+
launcher is able to correctly play the example streams from
http://opus-codec.org/examples

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:

LayoutTests:

Opus (http://opus-codec.org) streams are typically embedded into
Ogg containers, so add a check for it in the test case.

* media/media-can-play-ogg-expected.txt:
* media/media-can-play-ogg.html:
* platform/chromium-android/media/media-can-play-ogg-expected.txt:
* platform/chromium/media/media-can-play-ogg-expected.txt: Copied from LayoutTests/media/media-can-play-ogg-expected.txt.
* platform/efl/media/media-can-play-ogg-expected.txt: Copied from LayoutTests/media/media-can-play-ogg-expected.txt.
* platform/gtk/media/media-can-play-ogg-expected.txt: Copied from LayoutTests/media/media-can-play-ogg-expected.txt.
* platform/qt-linux/media/media-can-play-ogg-expected.txt: Copied from LayoutTests/media/media-can-play-ogg-expected.txt.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@140532 268f45cc-cd09-0410-ab3c-d52691b4dbfc