Some formerly-fixed objects scroll as if they are still fixed https://bugs.webkit.org/show_bug.cgi?id=107869
Reviewed by Simon Fraser.

Removing nodes need to mark m_hasChangedProperties on the state tree, otherwise 
they will not be committed to the scrolling tree until something else changes, 
which could actually take a while. During the delay, the scrolling tree will 
continue to move around the formerly-fixed object as long as it still has a layer.
* page/scrolling/ScrollingStateTree.cpp:
(WebCore::ScrollingStateTree::didRemoveNode):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@140747 268f45cc-cd09-0410-ab3c-d52691b4dbfc