REGRESSION (r141281): Navigating to this HTTP Live Streaming (application/vnd.apple.mpegurl) URL downloads a file instead of playing it in the browser https://bugs.webkit.org/show_bug.cgi?id=108674 <rdar://problem/13133595>
Reviewed by Eric Carlson.

When passed a HashSet of available MIME types to fill out, QTKit was overwriting the contents, which were
AVFoundation's supported types. Instead, it should be appending to the list of supported types.

* platform/graphics/mac/MediaPlayerPrivateQTKit.mm:
(WebCore::concatenateHashSets): Added static utility function.
(WebCore::MediaPlayerPrivateQTKit::getSupportedTypes): Concatenate instead of overwriting the passed in MIME types.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@141630 268f45cc-cd09-0410-ab3c-d52691b4dbfc