[GTK] Building fails in an armv5tel board https://bugs.webkit.org/show_bug.cgi?id=108083
Patch by Adrian Perez de Castro <aperez@igalia.com> on 2013-02-01
Reviewed by Martin Robinson.

In some systems in which $architecture is filled-in from "uname -m"
the value may not start with "arm-", but with "armvN", where "N" is
an architecture version number. The regexp in isARM() is modified
so it covers these kind of cases.

* Scripts/webkitdirs.pm:
(isARM):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@141683 268f45cc-cd09-0410-ab3c-d52691b4dbfc