[Chromium] Split calendar-picker-key-operations.html into two https://bugs.webkit.org/show_bug.cgi?id=109006
Reviewed by Kentaro Hara.

Split calendar-picker-key-operations.html into two parts:
 - OS-independent part (calendar-picker-key-operations.html), and
 - OS-dependent part (calendar-picker-f4-key.html).

We had some troubles when we updated calendar-picker-key-operations.html
because the behavior by F4 key is OS-dependent. We move the test for F4
key to new test.

* platform/chromium-mac/platform/chromium/fast/forms/calendar-picker/calendar-picker-f4-key-expected.txt:
Added. It is expected that this contains FAIL line because we don't
support the F4 key behavior on OSX.
* platform/chromium-win/platform/chromium/fast/forms/calendar-picker/calendar-picker-key-operations-expected.txt:
Removed. Now the result of calendar-picker-key-operations.html is OS-independent.
* platform/chromium/fast/forms/calendar-picker/calendar-picker-f4-key-expected.txt: Added.
* platform/chromium/fast/forms/calendar-picker/calendar-picker-f4-key.html:
Added. Move from calendar-picker-key-operations.html.
* platform/chromium/fast/forms/calendar-picker/calendar-picker-key-operations.html:
Move the F4 key part to calendar-picker-f4-key.html.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@141968 268f45cc-cd09-0410-ab3c-d52691b4dbfc