Load event fires too early with threaded HTML parser (take 2) https://bugs.webkit.org/show_bug.cgi?id=109485
Reviewed by Eric Seidel.

Source/WebCore:

This patch restores the code that was removed in
http://trac.webkit.org/changeset/142492 and adds code to
DocumentLoader.cpp to avoid the regression.

* dom/Document.cpp:
(WebCore::Document::hasActiveParser):
(WebCore::Document::decrementActiveParserCount):
* loader/DocumentLoader.cpp:
(WebCore::DocumentLoader::isLoadingInAPISense):

LayoutTests:

This patch also fixes a bug whereby removing an iframe during the load
event would trigger DumpRenderTree to dump the test in the middle of
the load event. We now wait until the load event is over.

* compositing/iframes/remove-iframe-crash-expected.txt:
* fast/frames/iframe-access-screen-of-deleted-expected.txt:
* fast/frames/remove-frame-during-load-event-expected.txt: Added.
* fast/frames/remove-frame-during-load-event.html: Added.
* http/tests/misc/xslt-bad-import-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@142555 268f45cc-cd09-0410-ab3c-d52691b4dbfc