WebVTTParser copies character buffer more often than necessary https://bugs.webkit.org/show_bug.cgi?id=103319
Reviewed by Eric Carlson.

Previously this codepath was creating a String just to parse the timestamp
and then if the timestamp was valid, was creating a second string.
I've fixed it to only create one string and use it in both places.
I also fixed this codepath to use 8bit strings when possible, per the FIXME.

* html/track/WebVTTParser.cpp:
(WebCore::WebVTTParser::constructTreeFromToken):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@143563 268f45cc-cd09-0410-ab3c-d52691b4dbfc